requirements.txt
pytest
pytest-dotenv
pytest-watch

icecream
pyvis
networkx
fastapi[standard]
jinja2
spacy[apple]